Removing contracts/web3 functions from InstantiableConfig class
